api/requirements.txt
alembic==1.10.3
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
cryptography==39.0.1
ecdsa==0.18.0
fastapi==0.92.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
mysql-connector-python==8.0.32
passlib==1.7.4
protobuf==3.20.3
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.5
PyJWT==2.6.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
starlette==0.25.0
typing_extensions==4.5.0
uvicorn==0.20.0
